Consolidate route restrictions into AuthRestriction
Replace DeveloperRestriction and SuperuserRestriction with a single
AuthRestriction module exposing ::Developer and ::Superuser, sharing the
cookie auth lookup. Merge their request specs to match.
